+++ deployment/v2/build-all.sh
@@ -20,6 +20,10 @@ echo "building..."
 time lerna run $bail --stream $runtarget || fail
 
 echo "============================================================================================="
+echo "running cfn-guard..."
+~/.guard/bin/cfn-guard validate -r ~/.guard/rules/aws-solutions.guard -d **/**/**/test/**/*.template.json
+
+echo "============================================================================================="
 echo "packaging..."
 time lerna run --bail --stream jsii-pacmak || fail